Revert "Adding IoT Hub WebSockets APIs, fixing Y2k38 issue (#924)"
This reverts commit 6a966901e467747d66e5705a3e25848b9ef45be8.

+++ sdk/inc/azure/iot/az_iot_provisioning_client.h
@@ -147,7 +147,7 @@ AZ_NODISCARD az_result az_iot_provisioning_client_get_client_id(
  */
 AZ_NODISCARD az_result az_iot_provisioning_client_sas_get_signature(
     az_iot_provisioning_client const* client,
-    uint64_t token_expiration_epoch_time,
+    uint32_t token_expiration_epoch_time,
     az_span signature,
     az_span* out_signature);
 
@@ -175,7 +175,7 @@ AZ_NODISCARD az_result az_iot_provisioning_client_sas_get_signature(
 AZ_NODISCARD az_result az_iot_provisioning_client_sas_get_password(
     az_iot_provisioning_client const* client,
     az_span base64_hmac_sha256_signature,
-    uint64_t token_expiration_epoch_time,
+    uint32_t token_expiration_epoch_time,
     az_span key_name,
     char* mqtt_password,
     size_t mqtt_password_size,
